Charge receipt email (#95)
* Adds support for the `receipt_email` parameter on Charge objects

The `receipt_email` parameter is the email address to send this
charge's receipt to. The receipt will not be sent until the charge
is paid. If this charge is for a customer, the email address
specified here will override the customer's email address.

Receipts will not be sent for test mode charges.

If `receipt_email` is specified for a charge in live mode, a receipt
will be sent regardless of your email settings.

Signed-off-by: Zach Wick <zwick@stripe.com>

* cherry-pick updates to _hash_to_object because unit tests fail on SubscriptionItem

* add receipt_email to form_fields for Net::Stripe::Charge

* add unit tests for receipt_email